pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "security-review-automation"
version = "0.1.0"
description = "AI-powered automated security code review using LLMs"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Alisid07", email = "ali.sarmad@gmail.com"}]

dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
    "gitpython>=3.1.40",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
security-review = "security_reviewer.cli:main"